@@ -15,11 +15,11 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.0",
-    "@docusaurus/plugin-ideal-image": "^2.4.0",
-    "@docusaurus/preset-classic": "^2.4.0",
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/plugin-ideal-image": "^3.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
     "@giscus/react": "^2.2.6",
-    "@mdx-js/react": "^1.6.22",
+    "@mdx-js/react": "^3.0.0",
     "clsx": "^1.2.1",
     "cookie": "^0.5.0",
     "docusaurus-plugin-image-zoom": "^0.1.1",
@@ -27,24 +27,24 @@
     "hast-util-is-element": "1.1.0",
     "js-yaml-loader": "^1.2.2",
     "medium-zoom": "^1.0.8",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-cookie-consent": "^8.0.1",
-    "react-dom": "^17.0.2",
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
+    "react-cookie-consent": "^9.0.0",
+    "react-dom": "^18.2.0",
     "react-loadable": "*",
     "react-player": "^2.11.0",
-    "react-share": "^4.4.1",
+    "react-share": "^5.0.0",
     "react-twitter-embed": "^4.0.4",
-    "remark-link-card": "^1.3.1",
     "sass": "^1.57.1"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.0",
+    "@docusaurus/module-type-aliases": "^3.0.0",
+    "@docusaurus/types": "3.0.0",
     "@tsconfig/docusaurus": "^1.0.6",
     "@types/js-yaml": "^4.0.5",
     "front-matter": "^4.0.2",
     "sass-loader": "^13.2.0",
-    "typescript": "^4.9.4",
+    "typescript": "^5.0.0",
     "webpack": "^5.75.0"
   },
   "browserslist": {
