@@ -63,4 +63,4 @@ const sidebars = {
   ],
 };
 
-module.exports = sidebars;
+export default sidebars;
